Move bolos api to bolo module and add heartbeat function
